+++ src/noahtheduke/splint.clj
@@ -40,6 +40,7 @@
    noahtheduke.splint.rules.lint.locking-object
    noahtheduke.splint.rules.lint.loop-do
    noahtheduke.splint.rules.lint.loop-empty-when
+   noahtheduke.splint.rules.lint.misplaced-type-hint
    noahtheduke.splint.rules.lint.missing-body-in-when
    noahtheduke.splint.rules.lint.not-empty
    noahtheduke.splint.rules.lint.prefer-method-values